Implement cache service metrics service
