+++ api/package.json
@@ -16,6 +16,7 @@
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
+    "ethers": "^5.6.4",
     "express": "^4.17.1",
     "express-rate-limit": "^5.5.0",
     "moment": "^2.29.1",
@@ -24,6 +25,7 @@
     "pg": "^8.6.0",
     "pg-format": "^1.0.4",
     "pg-native": "^3.0.0",
+    "prom-client": "^14.0.1",
     "web3": "^1.6.1"
   }
 }
