+++ api/src/config/index.js
@@ -1,0 +1,8 @@
+const { blockchains } = require('./blockchains');
+const cereTypes = require('./cereTypes');
+
+module.exports = { 
+  cereTypes, 
+  blockchains,
+  cacheIntervalMs: 10 * 60 * 1000
+};
